move commands into folders based on category
