+++ src/Discord/Commands/admin/restart.js
