+++ src/Discord/Commands/info/ping.js
