+++ src/Discord/Commands/info/serverInfo.js
