+++ src/Discord/Commands/manager/commands.js
