+++ src/Discord/Commands/misc/define.js
