+++ src/Discord/Commands/moderation/ban.js
@@ -1,5 +1,5 @@
 const { MessageEmbed } = require("discord.js");
-const { resolveUser, formatFromNow } = require("../../utils/functions");
+const { resolveUser, formatFromNow } = require("../../../utils/functions");
 
 module.exports = {
 	name: "ban",
